Restructureed seasonalStatAdapter into the mongoDbAdapter
