+++ src/sockets/commands/admin/acreatestat.ts
@@ -2,29 +2,28 @@
 
 import { Command } from '../types';
 import { SocketUser } from '../../types';
-import seasonAdapter from '../../../databaseAdapters/season';
 import { sendReplyToCommand } from '../../sockets';
-import { ISeason } from '../../../models/types/season.types';
 import { IUser } from '../../../gameplay/types';
-import seasonalStatAdapter from '../../../databaseAdapters/seasonalStat';
 import { ISeasonalStat } from '../../../models/types/seasonalStats.types';
+import mongoDbAdapter from '../../../databaseAdapters/mongoose';
+import { ISeason } from '../../../models/types/season';
 
 export const agetstat: Command = {
   command: 'ags',
   help: '/ags: Get stat',
   run: async (args: string[], socket: SocketUser) => {
-    const season: ISeason = await seasonAdapter.getCurrentSeason();
+    const season: ISeason = await mongoDbAdapter.season.getCurrentSeason();
     const user: IUser = socket.request.user;
 
-    const stat: ISeasonalStat = await seasonalStatAdapter.getStat(
+    const stat: ISeasonalStat = await mongoDbAdapter.seasonalStat.getStat(
       user._id,
-      season._id,
+      season.id,
     );
 
     const message = `
       Current stat for ${
         user.username
-      }: ${seasonalStatAdapter.parseSeasonalStat(stat)}`;
+      }: ${mongoDbAdapter.seasonalStat.parseSeasonalStat(stat)}`;
 
     sendReplyToCommand(socket, message);
   },
@@ -38,31 +37,31 @@ export const aupdatestat: Command = {
       sendReplyToCommand(socket, 'State "win" or "lose"');
     }
 
-    const season: ISeason = await seasonAdapter.getCurrentSeason();
+    const season: ISeason = await mongoDbAdapter.season.getCurrentSeason();
     const user: IUser = socket.request.user;
 
     let stat: ISeasonalStat;
 
     if (args[1] === 'win') {
-      stat = await seasonalStatAdapter.updateStat(
+      stat = await mongoDbAdapter.seasonalStat.updateStat(
         user._id,
-        season._id,
+        season.id,
         true,
         25,
       );
     } else {
-      stat = await seasonalStatAdapter.updateStat(
+      stat = await mongoDbAdapter.seasonalStat.updateStat(
         user._id,
-        season._id,
+        season.id,
         false,
         -10,
       );
     }
 
     const message = `
-      New stat for ${user.username}: ${seasonalStatAdapter.parseSeasonalStat(
-      stat,
-    )}`;
+      New stat for ${
+        user.username
+      }: ${mongoDbAdapter.seasonalStat.parseSeasonalStat(stat)}`;
 
     sendReplyToCommand(socket, message);
   },
